Updated flow, added actual functionality of allow comments in view

+++ src/app/admin/all-post/all-post.component.ts
@@ -33,8 +33,15 @@ export class AllPostComponent implements OnInit {
   }
 
   deletePost(data){
-    this.postService.deletePost(data._id).then(()=>{
-      this.fetchAllPosts();
-    })
+    let confirmRespose = confirm("Are you sure you want to delete this blog post? You cannot undo this action later.")
+    if(confirmRespose){
+      this.postService.deletePost(data._id).then(()=>{
+        this.fetchAllPosts();
+      })
+    }
+  }
+
+  viewComments(data){
+    this.router.navigate(['/admin','comments',data._id])
   }
 }
